Extra scene work.
The scene now contains 4 buttons, of which one is blinking.
For this to work I had to change the timer in Spaceship from a long to either a float or a double (i chose float, because of repetetive addition), because long lacks the precision to keep track of the updates and the timer stays 0.