Add Smoke Test.
Although this smoke test works fine on my machine, I can imagine it might
lead to some trouble on the CI server. In addition it unfortunately does
not increase the test coverage in Cobertura, probably because all the
production code is executed in a separate thread (which is necessary to be
able to kill the application). Therefore I ignored the test for now.