Add intensity update mechanism.
Two constructors for the EventScheduler exits now. One to create a fixed
intensity scheduler, and one to create a variable intensity scheduler.
When updating the fixed intensity scheduler with a given percentage,
nothing happens. When updating a variable intensity scheduler with a given
percentage, it linearly interpolates between its predefined minimum and
maximum intensities. Updated test suite to reflect the changes.