Added more tests to increase coverage
There are two missing branches in the RouteGenerator and PlayState class
which aren't covered yet.
The reason why is because the RouteGenerator uses a random generator in
the middle of the method that is difficult to mock, while PlayState uses
the VRApplication, which causes a nullpointerException if I try to use
it.
Overall, the branch coverage is now around 79.8%