Refactor Event Factory.
Refactor the Event Factory to a set of new components. Apologies for the
large commit, but in order to implement the Abstract Factory and the
Builder design pattern I had to make changes all over the place to keeps
things working. Now that the implementation is refactored, taking into
account also the feedback from our Software Engineering TA, finally the
Event Parameters can be integrated. Improving test coverage and fixing
static errors will be done in upcoming commits.